requirements.txt
# Core
streamlit==1.37.0
streamlit-autorefresh==1.0.1
streamlit-aggrid==0.3.4

# Data & ML
numpy==1.26.4
pandas==2.3.3
scikit-learn==1.7.2
joblib==1.5.2

# Visualization
plotly==5.23.0

# Async & Networking
aiohttp==3.9.5
requests==2.32.3
tenacity==8.2.2

# Threat Intelligence
virustotal-python==1.0.2
vt-py==0.22.0

# Security / Analysis
yara-python==4.5.0
python-magic-bin==0.4.14
pefile==2024.8.26
psutil==7.1.3
cryptography==45.0.6

# GeoIP
geoip2==4.8.1
maxminddb==2.6.2

# Utilities
python-dotenv==1.2.1
click==8.1.7
watchdog==4.0.1
protobuf==4.25.3
packaging==24.1
typing-extensions>=4.8.0,<5

# Build
Nuitka>=2.8.9
